+++ Player.cpp
@@ -55,11 +55,11 @@ void Player::remove_card(int val) {
     num_card--;
 }
 
-int Player::make_move(const Board &board, const Player &human, const Player &ai, int &row, int &col, int &val) {
+float Player::make_move(const Board &board, const Player &human, const Player &ai, int &row, int &col, int &val) {
     Node root(board, human, ai);
     int max_depth = estimate_max_depth(board, human, ai);
     debug(max_depth);
-    int max_score = root.search_ab(0, max_depth, false, MIN, MAX);
+    float max_score = root.search_ab(0, max_depth, false, MIN, MAX);
     debug(max_score);
     root.get_best_move(row, col, val);
